house_price_prediction/server/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
pymongo
python-dotenv
joblib